Changes order tiles are rendered.
The tiles are rendered in a different order so that the perceived speed
of the application is improved. This switches the tiles around so that
the visible layer is fetched and drawn before the cached layers, on a
tile by tile basis. However, there are still more changes that need to
be done to make sure that all the visible tiles are rendered and fetched
before the cached layers.