Fixing bug: layer change not reflected in React.
When using the keyboard to change the layer in the tile viewer, the
changes were not being reflected in the React control components. This
resulted in the tileviewer jumping around after scrolling and reverting
to the original layer that was set before the keyboard changes were
made. The fix is to pass the react component to the tileviewer code and
have it update the state when the tileviewer registers a keypress. This
tightly couples the openseadragon code with the tileviewer, but I think
that ship has already sailed.